+++ src/main/java/com/steadybit/mcp/Tools.java
@@ -1,0 +1,8 @@
+/*
+ * Copyright 2025 steadybit GmbH. All rights reserved.
+ */
+
+package com.steadybit.mcp;
+
+public interface Tools {
+}