+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 
-## next
+## v0.0.7
 
+- Updated dependencies to latest versions, fixing some CVE issues
 - Removed `get_api_spec` tool, content gets truncated, and it's not really MCP agnostic / helpful.
+- Added `list_environments` tool to retrieve environments
+- Added `list_teams` tool to retrieve teams
+- Added `list_experiment_schedules` tool to retrieve experiment schedules
+- Added `list_experiment_templates` and `get_experiment_template` tool to interact with templates.
+- Added `create_experiment_from_template` tool to create experiments from templates. You explicitly need to enable this
+  capability via an environment variable `CAPABILITIES_ENABLED_0=CREATE_EXPERIMENT`.
 
 ## v0.0.6